Moved to TS and improved Speed metrics
All Done
