+++ .eslintrc.json
@@ -1,3 +1,10 @@
 {
-  "extends": "next/core-web-vitals"
+  "extends": ["next", "next/core-web-vitals", "eslint:recommended"],
+  "globals": {
+    "React": "readonly",
+    "NodeJS": true
+  },
+  "rules": {
+    "@next/next/no-img-element": "off"
+  }
 }
