+++ package.json
@@ -1,32 +1,38 @@
 {
   "name": "heroicons.com",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "format": "prettier --cache --write ."
   },
   "dependencies": {
+    "@fontsource/inter": "^4.5.12",
     "@headlessui/react": "^1.6.6",
-    "autoprefixer": "^10.4.8",
     "clsx": "^1.2.1",
-    "focus-visible": "^5.2.0",
-    "heroicons": "^2.0.10",
     "match-sorter": "^6.3.1",
     "next": "12.2.5",
-    "postcss": "^8.4.16",
-    "postcss-focus-visible": "^7.1.0",
-    "raw-loader": "^4.0.2",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwindcss": "^3.1.8"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "eslint": "8.22.0",
+    "@types/node": "18.7.15",
+    "@types/react": "18.0.18",
+    "@types/react-dom": "18.0.6",
+    "@types/webpack-env": "^1.18.0",
+    "autoprefixer": "^10.4.8",
+    "eslint": "8.23.0",
     "eslint-config-next": "12.2.5",
+    "focus-visible": "^5.2.0",
+    "heroicons": "^2.0.10",
+    "postcss": "^8.4.16",
+    "postcss-focus-visible": "^7.1.0",
     "prettier": "^2.7.1",
-    "prettier-plugin-tailwindcss": "^0.1.13"
+    "prettier-plugin-tailwindcss": "^0.1.13",
+    "tailwindcss": "^3.1.8",
+    "typescript": "4.8.2"
   }
 }
