+++ src/pages/_app.tsx
@@ -1,27 +1,29 @@
-import Head from 'next/head'
 import '@/styles/tailwind.css'
+import '@fontsource/inter/variable.css'
 import 'focus-visible'
+import type { AppProps } from 'next/app'
+import Head from 'next/head'
 
-function MyApp({ Component, pageProps }) {
+export default function MyApp({ Component, pageProps }: AppProps) {
   return (
     <>
       <Head>
         <link
           rel="apple-touch-icon"
           sizes="180x180"
-          href={require('@/images/apple-touch-icon.png').default.src}
+          href="/apple-touch-icon.png"
         />
         <link
           rel="icon"
           type="image/png"
           sizes="32x32"
-          href={require('@/images/favicon-32x32.png').default.src}
+          href="/favicon-32x32.png"
         />
         <link
           rel="icon"
           type="image/png"
           sizes="16x16"
-          href={require('@/images/favicon-16x16.png').default.src}
+          href="/favicon-16x16.png"
         />
         <meta
           content="Beautiful hand-crafted SVG icons, by the makers of Tailwind CSS."
@@ -36,10 +38,7 @@ function MyApp({ Component, pageProps }) {
         />
         <title>Heroicons</title>
         <meta property="og:title" content="Heroicons" />
-        <meta
-          property="og:image"
-          content={`https://heroicons.com${require('@/images/social-card.jpg').default.src}`}
-        />
+        <meta property="og:image" content="/social-card.jpg" />
         <meta name="twitter:card" content="summary_large_image" />
         <meta name="twitter:site" content="@tailwindlabs" />
         <meta name="twitter:title" content="Heroicons" />
@@ -47,14 +46,9 @@ function MyApp({ Component, pageProps }) {
           name="twitter:description"
           content="Beautiful hand-crafted SVG icons, by the makers of Tailwind CSS."
         />
-        <meta
-          name="twitter:image"
-          content={`https://heroicons.com${require('@/images/social-card.jpg').default.src}`}
-        />
+        <meta name="twitter:image" content="social-card.jpg" />
       </Head>
       <Component {...pageProps} />
     </>
   )
 }
-
-export default MyApp
