+++ src/types/index.ts
@@ -1,0 +1,7 @@
+export interface IconProps {
+  name: string
+  tags: string[]
+  svg: SVGSVGElement | string | any
+}
+
+export type IconTypes = 'Mini' | 'Outline' | 'Solid'
