+++ v1/next.config.js
@@ -15,13 +15,18 @@ module.exports = {
               .replace(/>(\s|\\n)+</g, '><')
               .replace(/\\n$/, '')
               .replace(/\\"/g, '"')
-              .replace(/(\s)([a-z-]+)="([^"]+)"/gi, (_, prefix, attr, value) => {
-                const jsxValue = /^[0-9.]+$/.test(value) ? `{${value}}` : `"${value}"`
-                return `${prefix}${attr.replace(
-                  /-([a-z])/gi,
-                  (_, letter) => `${letter.toUpperCase()}`
-                )}=${jsxValue}`
-              })
+              .replace(
+                /(\s)([a-z-]+)="([^"]+)"/gi,
+                (_, prefix, attr, value) => {
+                  const jsxValue = /^[0-9.]+$/.test(value)
+                    ? `{${value}}`
+                    : `"${value}"`
+                  return `${prefix}${attr.replace(
+                    /-([a-z])/gi,
+                    (_, letter) => `${letter.toUpperCase()}`
+                  )}=${jsxValue}`
+                }
+              )
               .replace(
                 /<svg[^>]+>(.*?)<\/svg>/s,
                 (svg.match(/</g) || []).length > 3 ? '(<>$1</>)' : '($1)'
