+++ v1/src/data/tags.js
@@ -19,8 +19,22 @@ export default {
   'chart-bar': ['graph', 'stats', 'statistics', 'analytics'],
   'chart-pie': ['graph', 'stats', 'statistics', 'analytics'],
   'chart-square-bar': ['graph', 'stats', 'statistics'],
-  'chat-alt-2': ['speech', 'bubble', 'comment', 'response', 'message', 'discussion'],
-  'chat-alt': ['speech', 'bubble', 'comment', 'response', 'message', 'discussion'],
+  'chat-alt-2': [
+    'speech',
+    'bubble',
+    'comment',
+    'response',
+    'message',
+    'discussion',
+  ],
+  'chat-alt': [
+    'speech',
+    'bubble',
+    'comment',
+    'response',
+    'message',
+    'discussion',
+  ],
   chat: ['speech', 'bubble', 'comment', 'response', 'message', 'discussion'],
   'check-circle': ['tick', 'correct', 'confirm'],
   check: ['tick', 'correct', 'valid', 'ok', 'confirm'],
